fix: resolve TypeScript errors in test files and source code
- Fix Cheerio mock typing by adding 'as any' type assertion to prevent 'html' method errors
- Fix optional chaining for potentially undefined properties in Mx.test.mts
- Add missing beforeEach import in Svg.test.mts
- Remove third parameter from cheerio.load() calls in C4 components (System, Container, Component, Relationship)
- Convert test files from .ts to .mts extension for consistency with source files
- Add type assertions for object array access to prevent TypeScript strict mode errors

All tests now pass with proper TypeScript compilation.
