+++ PR_DETAILS.md
@@ -1,0 +1,14 @@
+# Pull Request Details
+
+## URL to Create PR
+https://github.com/localgod/catalyst/compare/main...feature/vitest-testing-setup?expand=1
+
+## Title
+feat: Add comprehensive Vitest testing setup with 97.62% coverage
+
+## Description
+
+### Summary
+
+This PR introduces a comprehensive testing setup using Vitest with extensive coverage for the Catalyst project.
+
